+++ chat-components/src/components/outofofficehourspane/common/presetFourProps/presetFourOOOHPaneProps.ts
@@ -1,0 +1,8 @@
+import { IOOOHPaneProps } from "../../interfaces/IOOOHPaneProps";
+import { defaultOOOHPaneStyles } from "../defaultProps/defaultStyles/defaultOOOHPaneStyles";
+import { presetFourOOOHPaneControlProps } from "./presetFourOOOHPaneControlProps";
+
+export const presetFourOOOHPaneProps: IOOOHPaneProps = {
+    controlProps: presetFourOOOHPaneControlProps,
+    styleProps: defaultOOOHPaneStyles
+};