+++ chat-widget/src/components/livechatwidget/common/defaultProps/dummyDefaultProps.ts
@@ -13,9 +13,9 @@ import { createActivityMiddleware } from "../../../webchatcontainerstateful/webc
 import createAttachmentMiddleware from "../../../webchatcontainerstateful/webchatcontroller/middlewares/renderingmiddlewares/attachmentMiddleware";
 import { createAvatarMiddleware } from "../../../webchatcontainerstateful/webchatcontroller/middlewares/renderingmiddlewares/avatarMiddleware";
 import { createMarkdown } from "../createMarkdown";
+import { createWebChatTelemetry } from "../../../webchatcontainerstateful/webchatcontroller/webchattelemetry/WebChatLogger";
 import { groupActivitiesMiddleware } from "../../../webchatcontainerstateful/webchatcontroller/middlewares/renderingmiddlewares/groupActivitiesMiddleware";
 import { typingIndicatorMiddleware } from "../../../webchatcontainerstateful/webchatcontroller/middlewares/renderingmiddlewares/typingIndicatorMiddleware";
-import { createWebChatTelemetry } from "../../../webchatcontainerstateful/webchatcontroller/webchattelemetry/WebChatLogger";
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export const dummyDefaultProps: ILiveChatWidgetProps = {
@@ -1113,7 +1113,7 @@ export const dummyDefaultProps: ILiveChatWidgetProps = {
             subtitleText: "Please take a moment to give us feedback about your chat experience. We are loading the survey for you now."
         },
         styleProps: {
-            // ...[Existing chat button style props]
+            // ...[Existing loading pane style props]
             generalStyleProps: {
                 position: "initial",
                 width: "100%",
@@ -1557,10 +1557,10 @@ export const dummyDefaultProps: ILiveChatWidgetProps = {
                 padding: "0px 10px 0 10px"
             },
             userMessageBoxStyles: {
-                maxWidth: "75%"
+                maxWidth: "90%"
             },
             systemMessageBoxStyles: {
-                maxWidth: "75%"
+                maxWidth: "90%"
             }, 
             typingIndicatorStyleProps: {
                 marginLeft: "10px",