+++ chat-widget/src/components/webchatcontainerstateful/webchatcontroller/middlewares/renderingmiddlewares/defaultStyles/defaultSystemMessageBoxStyles.ts
@@ -1,3 +1,3 @@
 export const defaultSystemMessageBoxStyles = {
-    maxWidth: "75%"
+    maxWidth: "90%"
 };